Introduce test environment
